% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_seconds.R
\name{bin_seconds}
\alias{bin_seconds}
\title{bin_seconds}
\usage{
bin_seconds(
  seconds.dir,
  output.dir,
  bin.to.mins,
  cross.recordings = TRUE,
  align.rows = c("midnight", "recorder-on"),
  roll.by.mins = NULL,
  subset.by.col = NULL,
  overwrite.existing = FALSE
)
}
\arguments{
\item{seconds.dir}{Directory (string) containing seconds CSV files.}

\item{output.dir}{Directory (string) to store binned CSV files}

\item{bin.to.mins}{Character or numeric. Time interval to bin to (num MINS or "Total")}

\item{cross.recordings}{Logical. Should bins be calculated across or within recordings? (default = TRUE)}

\item{align.rows}{Character. opts = c("midnight", "recorder-on") Should the first bin start at midnight (00:00:00) the day the recorder was turned on ("midnight") or the first second in which the recorder was on ("recorder-on")}

\item{roll.by.mins}{Numeric. Default = NULL (sequential bins, eg 0-59, 60-119) How many minutes between bins (or, do bins overlap?). Ex, if bin.to.mins = 60 & roll.by.mins = 1, then 1st bin = minutes 0-59, 2nd bin = mins 1-60, 3rd = mins 2-61, etc.}

\item{subset.by.col}{Character. Include only rows where this column is NOT 0 (default = NULL)}

\item{overwrite.existing}{Logical. If the target file already exists, should the function overwrite it? If FALSE (default), subject will be skipped if file exists. If TRUE, new file will ovewrite existing file.}
}
\value{
A dataframe and/or CSV file.
}
\description{
This function imports standardized secMidnight files and bins them to specified minutes, starting from midnight.
}
\examples{
bin_seconds(
seconds.dir = "Server:/LENAData/secMidnight",
output.dir = "Server:/LENAData/min1Midnight",
bin.to.mins = 5,
align.rows = "midnight")

}
