% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ITS_to_seconds.R
\name{ITS_to_seconds}
\alias{ITS_to_seconds}
\title{ITS_to_seconds}
\usage{
ITS_to_seconds(
  ITS.dir,
  CSV.dir,
  time.zone = NULL,
  write.recordings = FALSE,
  write.blocks = FALSE,
  write.segments = FALSE,
  write.centiseconds = FALSE,
  write.seconds = TRUE
)
}
\arguments{
\item{ITS.dir}{Directory (string) containing ITS files. Default = working
directory.}

\item{CSV.dir}{Directory (string) to store CSV files}

\item{time.zone}{OS-specific character string for time zone. To use current
system timezone, type "Sys.timezone()". For other options, run OlsonNames()
for list.}
}
\value{
4 CSVs per file input (recordings, blocks, segments, validation).
}
\description{
This function imports Recordings, Blocks, and Segments
  (uninterrupted human-created sounds, electronic noise, or silence) from an
  ITS file, converts them into a dataframe, then adds time from desired time
  zone in clocktime and seconds from midnight.
}
\examples{
ITS_to_seconds(
ITS.dir = "SERVER:/ITS_Files/",
CSV.dir = "SERVER:/CSVOutput/",
time.zone = "America/Los_Angeles")
}
