% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_recordings.R
\name{remove_recordings}
\alias{remove_recordings}
\title{remove_recordings

This function removes specified recordings from an ITS file, adjusts the recording start and end centisecond counts, then saves a NEW ITS file with the name "originalname_edited.its".

For more information, including example outputs and explanations of all output columns, see \url{https://htanderson.github.io/ITSbin/}}
\usage{
remove_recordings(
  ITS.folder,
  ITS.file,
  edited.ITS.folder,
  recordings.to.remove
)
}
\arguments{
\item{ITS.folder}{Directory (string) containing ITS files.}

\item{ITS.file}{Single .its file to be altered (string)}

\item{edited.ITS.folder}{Directory (string) to store edited ITS file.}

\item{recordings.to.remove}{Integer: Recording number(s) to be removed?}
}
\value{
An its/xml file
}
\description{
remove_recordings

This function removes specified recordings from an ITS file, adjusts the recording start and end centisecond counts, then saves a NEW ITS file with the name "originalname_edited.its".

For more information, including example outputs and explanations of all output columns, see \url{https://htanderson.github.io/ITSbin/}
}
\examples{
\dontrun{

remove_recordings(
ITS.folder = "SERVER/ITS_Files/ITS",
ITS.file = "subj001.its",
recordings.to.remove = 1)

}

}
