% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeRecds.R
\name{removeRecds}
\alias{removeRecds}
\title{removeRecds}
\usage{
removeRecds(path.in = paste0(getwd(), "/"), ITS, removeNum, beforeRec)
}
\arguments{
\item{path.in}{Directory (string) containing ITS files. Default = working directory.}

\item{ITS}{.its file to be altered}

\item{removeNum}{Integer: Recording number(s) to be removed?}

\item{beforeRec}{Logical: Is/are the recording(s) to be removed before the recording(s) to be kept?}
}
\value{
An its/xml file
}
\description{
This function removes unwanted recordings from an ITS file and adjusts the recording start and end centisecond counts.
}
\examples{
removeRecds(path.in = "D:/SkyDrive/Documents/UOregon/16_FYP_LENA/ITS_Files/ITS", ITS = "EH_044_02.its", removeNum = 1, beforeRec = TRUE)
}
