% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_recordings.R
\name{remove_recordings}
\alias{remove_recordings}
\title{remove_recordings}
\usage{
remove_recordings(ITS.folder, ITS.file, remove.num, before.keep)
}
\arguments{
\item{ITS.folder}{Directory (string) containing ITS files.}

\item{ITS.file}{Single .its file to be altered}

\item{remove.num}{Integer: Recording number(s) to be removed?}

\item{before.keep}{Logical: Is/are the recording(s) to be removed before the recording(s) to be kept?}
}
\value{
An its/xml file
}
\description{
This function removes unwanted recordings from an ITS file and adjusts the recording start and end centisecond counts.
}
\examples{
removeRecds(
ITS.folder = "D:/SkyDrive/Documents/UOregon/16_FYP_LENA/ITS_Files/ITS",
ITS.file = "subj001.its",
remove.num = 1,
before.keep = TRUE)
}
